add monta env

+++ LSSutils/dataviz.py
@@ -254,7 +254,7 @@ def mollview(m, vmin, vmax, unit, use_mask=False,
     '''    
     nside     = hp.npix2nside(len(m))
     rotatedeg = np.degrees(rotate)    
-    ysize     = xsize/2.                       # ratio is always 1/2
+    ysize     = int(xsize/2.)                     # ratio is always 1/2
 
     
     # galactic plane
